+++ driver-core/src/main/java/com/datastax/driver/core/DefaultResultSetFuture.java
@@ -71,29 +71,59 @@ class DefaultResultSetFuture extends AbstractFuture<ResultSet> implements Result
                             ResultSet rs = ArrayBackedResultSet.fromMessage(rm, session, protocolVersion, info, statement);
                             switch (scc.change) {
                                 case CREATED:
-                                    if (scc.name.isEmpty()) {
-                                        session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, null, null);
-                                    } else {
-                                        session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null);
+                                    switch (scc.target) {
+                                        case KEYSPACE:
+                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null, null);
+                                            break;
+                                        case TABLE:
+                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, scc.name, null);
+                                            break;
+                                        case TYPE:
+                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null, scc.name);
+                                            break;
                                     }
                                     break;
                                 case DROPPED:
-                                    if (scc.name.isEmpty()) {
-                                        // If that the one keyspace we are logged in, reset to null (it shouldn't really happen but ...)
-                                        // Note: Actually, Cassandra doesn't do that so we don't either as this could confuse prepared statements.
-                                        // We'll add it back if CASSANDRA-5358 changes that behavior
-                                        //if (scc.keyspace.equals(session.poolsState.keyspace))
-                                        //    session.poolsState.setKeyspace(null);
-                                        session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, null, null);
-                                    } else {
-                                        session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null);
+                                    KeyspaceMetadata keyspace;
+                                    switch (scc.target) {
+                                        case KEYSPACE:
+                                            // If that the one keyspace we are logged in, reset to null (it shouldn't really happen but ...)
+                                            // Note: Actually, Cassandra doesn't do that so we don't either as this could confuse prepared statements.
+                                            // We'll add it back if CASSANDRA-5358 changes that behavior
+                                            //if (scc.keyspace.equals(session.poolsState.keyspace))
+                                            //    session.poolsState.setKeyspace(null);
+                                            session.cluster.manager.metadata.removeKeyspace(scc.keyspace);
+                                            break;
+                                        case TABLE:
+                                            keyspace = session.cluster.manager.metadata.getKeyspace(scc.keyspace);
+                                            if (keyspace == null)
+                                                logger.warn("Received a DROPPED notification for table {}.{}, but this keyspace is unknown in our metadata",
+                                                    scc.keyspace, scc.name);
+                                            else
+                                                keyspace.removeTable(scc.name);
+                                            break;
+                                        case TYPE:
+                                            keyspace = session.cluster.manager.metadata.getKeyspace(scc.keyspace);
+                                            if (keyspace == null)
+                                                logger.warn("Received a DROPPED notification for UDT {}.{}, but this keyspace is unknown in our metadata",
+                                                    scc.keyspace, scc.name);
+                                            else
+                                                keyspace.removeUserType(scc.name);
+                                            break;
                                     }
+                                    this.setResult(rs);
                                     break;
                                 case UPDATED:
-                                    if (scc.name.isEmpty()) {
-                                        session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null);
-                                    } else {
-                                        session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, scc.name);
+                                    switch (scc.target) {
+                                        case KEYSPACE:
+                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null, null);
+                                            break;
+                                        case TABLE:
+                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, scc.name, null);
+                                            break;
+                                        case TYPE:
+                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null, scc.name);
+                                            break;
                                     }
                                     break;
                                 default: